requirements.txt
RPi.GPIO
pygame
max7219
picamera
